add exception, refactoring
