+++ src/main/java/com/rybka/Application.java
@@ -3,7 +3,7 @@ package com.rybka;
 import com.rybka.dao.HibernateDAO;
 import com.rybka.service.BaseCurrencyExchangeConnector;
 import com.rybka.service.ExchangeService;
-import com.rybka.service.ExchangeRateBaseCurrencyExchangeConnector;
+import com.rybka.service.ExchangeRateConnector;
 import com.rybka.view.ExchangeView;
 import coresearch.cvurl.io.request.CVurl;
 
@@ -12,7 +12,7 @@ import java.util.Scanner;
 public class Application {
     public static void main(String[] args) {
 
-        BaseCurrencyExchangeConnector baseCurrencyExchangeConnector = new ExchangeRateBaseCurrencyExchangeConnector(new CVurl());
+        BaseCurrencyExchangeConnector baseCurrencyExchangeConnector = new ExchangeRateConnector(new CVurl());
 
         ExchangeView view = new ExchangeView(
                 new Scanner(System.in),
